fix: remove cycling dependency on types
